Refactor add transaction form
